% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gini_impurity}
\alias{gini_impurity}
\title{Gini impurity of linear split}
\usage{
gini_impurity(dat, label, par)
}
\arguments{
\item{dat}{Data frame of predictors.}

\item{label}{Vector of labels.}

\item{par}{Vector of coefficients of split.}
}
\value{
Gini impurity.
}
\description{
Compute Gini impurity of linear split.
}
\author{
Johannes Tillil
}
