% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gini_impurity_batch}
\alias{gini_impurity_batch}
\title{Batchwise gini impurity of candidate linear splits.}
\usage{
gini_impurity_batch(dat, label, candidates, varID)
}
\arguments{
\item{dat}{Data frame of predictors.}

\item{label}{Vector of labels.}

\item{par}{Vector of coefficients of split.}
}
\value{
Gini impurities.
}
\description{
Compute Batchwise gini impurity of candidate linear splits.
}
\author{
Johannes Tillil
}
