% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRFNetwork.R
\name{simpleRFNetwork}
\alias{simpleRFNetwork}
\title{simpleRFNetwork}
\usage{
simpleRFNetwork(
  formula,
  data,
  num_trees = 50,
  mtry = NULL,
  min_node_size = NULL,
  replace = TRUE,
  probability = FALSE,
  splitrule = NULL,
  unordered_factors = "ignore",
  num_threads = 1,
  varclusters = NULL,
  splitobject = NULL,
  splitmethod = NULL,
  varselection = NULL,
  seed = 1L
)
}
\arguments{
\item{formula}{Object of class \code{formula} or \code{character} describing the model to fit.}

\item{data}{Training data of class \code{data.frame}.}

\item{num_trees}{Number of trees.}

\item{mtry}{Number of variables to possibly split at in each node.}

\item{min_node_size}{Minimal node size. Default 1 for classification, 5 for regression, 3 for survival and 10 for probability estimation.}

\item{replace}{Sample with replacement. Default TRUE.}

\item{probability}{Grow a probability forest. Default FALSE.}

\item{splitrule}{Splitrule to use in trees. Default "Gini" for classification and probability forests, "Variance" for regression forests and "Logrank" for survival forests.}

\item{unordered_factors}{How to handle unordered factor variables. One of "ignore", "order_once", "order_split" and "partition" with default "ignore".}

\item{num_threads}{Number of threads used for mclapply, set to 1 for debugging.}

\item{varclusters}{List of numeric vectors that contain the IDs of the nodes in the respective module.}

\item{splitobject}{Object for node splitting. Use "single_variable" for standard procedure and "module" for use of complete modules in splitting.}

\item{splitmethod}{Method for node splitting. Use "SVM_linear" for a linear SVM line separating classes.}

\item{varselection}{Variable selection in case multiple variables are used for node splitting. Default "none" uses all variables of a module.}
}
\description{
Implements Random Forests (Breiman 2001) with emphasis on simplicity. 
Uses reference classes and only plain \code{R}. 
Not optimized for computation speed. 
Allows rapid prototyping of RF-type algorithms.
}
\details{
Unordered factor variables can be handled in different ways. 
Use "ignore" to treat them as ordered in the order of the factor levels. 
With "order_once" and "order_split" they are ordered by their response values. For "order_once" this is done once before the analysis, for "order_split" this is done in each split.
With "partition" all 2-partitions of the factor levels are considered for splitting.
}
\examples{
\donttest{
library(simpleRFNetwork) 

# Network Data Generation

testdat <- genGeneNetworkData(
  num_networks=1,
  num_genes=100,
  num_modules=NULL,
  num_observations=1000,
  num_causal_modules=1,
  num_causal_genes=3,
  effect_measure=0.1
)

# Create Random Forest

rf <- simpleRFNetwork(pheno ~ .,
                      data=testdat[[1]]$data,
                      num_trees=2,
                      num_threads=7,
                      splitobject="module",
                      splitmethod="LDA",
# alternative splitmethods: univariate_fast, CART_fast, LDA, SVM, Gini_optimal, Gini_stoch_optimal
                      varselection="none",
                      varclusters=testdat[[1]]$modules)

# Forest Diagnostics

rf$forest_time

rf$trees[[1]]$varclusters
rf$trees[[1]]$split_clusterIDs

rf$trees[[1]]$split_coefficients
rf$trees[[1]]$split_values

rf$trees[[1]]$linearcomb_times
rf$trees[[1]]$node_times
rf$trees[[1]]$sizes
rf$trees[[1]]$depths

rf$variableImportance(num_threads = 7)
rf$predictionError()

# TODO Prediction

train_idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris_train <- iris[train_idx, ]
iris_test <- iris[-train_idx, ]
rf_iris <- simpleRF(Species ~ ., data = iris_train)
pred_iris <- rf_iris$predict(iris_test)
table(iris_test$Species, pred_iris)
}

}
\references{
Breiman, L. (2001). Random forests. Mach Learn, 45(1), 5-32. \cr
}
\author{
Marvin N. Wright, Johannes Tillil
}
