% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeVarClusters.R
\docType{class}
\name{TreeVarClusters-class}
\alias{TreeVarClusters-class}
\alias{TreeVarClusters}
\title{Tree class}
\description{
Virtual class for Random forest tree where split variables are clusters. 
Contains all fields and methods used in all "VarClusters" tree subclasses.
}

