% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeVarClustersClassification.R
\docType{class}
\name{TreeVarClustersClassification-class}
\alias{TreeVarClustersClassification-class}
\alias{TreeVarClustersClassification}
\title{Classification tree class}
\description{
Subclass for classification tree where split variables are clusters.
Contains all fields and methods used special for classification trees.
}

